refactor(ui): move ViewRouter-rendered views to components/views/
TableView, ApplicationDashboard, ApplicationCatalogueView,
TechnologyCatalogueView and ElementCatalogueView are not routes —
they are components rendered by ViewRouter. Move them to
components/views/ to reflect that distinction.

routes/ now contains only components registered directly in App.svelte:
Home, WorkspaceOverview, ViewRouter, DependencyGraphView,
CapabilityTree, ApplicationLandscapeMap.
