+++ cmd/archipulse/ui/src/components/views/ApplicationDashboard.svelte
@@ -1,6 +1,6 @@
 <script>
   import { onMount } from 'svelte';
-  import { api } from '../lib/api.js';
+  import { api } from '../../lib/api.js';
   import { ArcChart } from 'layerchart';
 
   export let params = {};
