+++ cmd/archipulse/ui/src/components/views/ElementCatalogueView.svelte
@@ -1,6 +1,6 @@
 <script>
   import { onMount } from 'svelte';
-  import { api } from '../lib/api.js';
+  import { api } from '../../lib/api.js';
 
   export let params = {};
   $: wsId = params.wsId;
