+++ cmd/archipulse/ui/src/components/views/TableView.svelte
@@ -1,7 +1,7 @@
 <script>
   import { onMount } from 'svelte';
-  import { api } from '../lib/api.js';
-  import { VIEWS } from '../lib/views.js';
+  import { api } from '../../lib/api.js';
+  import { VIEWS } from '../../lib/views.js';
   import * as Table from '$lib/components/ui/table';
   import { Button } from '$lib/components/ui/button';
   import { Badge } from '$lib/components/ui/badge';
