+++ memory/MEMORY.md
@@ -1,0 +1,6 @@
+# ArchiPulse Memory
+
+- [ArchiPulse project status](project_status.md) — v0.5 in progress; current stack and completed work
+- [ArchiPulse technical decisions](project_tech.md) — stack, gotchas, conventions, local DB config
+- [No Claude attribution](feedback_no_attribution.md) — never add Co-Authored-By or Generated with Claude Code
+- [Git workflow](feedback_git_workflow.md) — always branch + PR, never push directly to main
