+++ memory/feedback_no_attribution.md
@@ -1,0 +1,11 @@
+---
+name: No Claude attribution in commits
+description: Never add Co-Authored-By Claude or Generated with Claude Code to commits or PRs
+type: feedback
+---
+
+Never add `Co-Authored-By: Claude` or "Generated with Claude Code" footers to commit messages or PR descriptions.
+
+**Why:** User preference — these are their commits and PRs.
+
+**How to apply:** Omit any Claude attribution from all git commits and PR bodies, always.
